added pokemon attacks

+++ Pokemon/source/src/sprite.cpp
@@ -25,8 +25,8 @@ Sprite::~Sprite() {}
 
 void Sprite::draw(Graphics &graphics, int x, int y) {
 	SDL_Rect destinationRectangle = { x, y,
-			this->_sourceRect.w * globals::SPRITE_SCALE,
-			this->_sourceRect.h * globals::SPRITE_SCALE};
+			this->_sourceRect.w * (int)globals::SPRITE_SCALE,
+			this->_sourceRect.h * (int)globals::SPRITE_SCALE};
 	graphics.blitSurface(this->_spriteSheet, &this->_sourceRect, &destinationRectangle);
 }
